Fix open file handles in TrueZIP
The old implementation of WarEngine#updateWarFile() does not close all open
file handles which could lead to out of memory errors.

The new implementation fixes the open file handles and removed the dependency
to TrueZIP. It uses JDK 7 feature ZipFileSystemProvider to append the lavender
files to the war file. It is slightly less performant compared to the old
implementation using TrueZIP (~40ms with TrueZIP vs ~140ms with
ZipFileSystemProvider).