Use production and development filter in test environment
The production filter rewrites the HTML output and replaces CDN
resources references. This filter is used in production.

The development filter resolves image paths to local, jar dependencies,
or remote paths, such as subversion urls. A remote resource is download
and served by the filter.

Up to now, the requirements were to serve either the production
(exclusive) or the development filter. New feature requrests make this
assumption void.

The new requirement is that in development environment the application
might use both, the HTML rewrite for external CDN resources and the
image resolving at the same time. Therefore, the development mode uses
the production filter as well.

The development mode is detected by the ServletContext parameter 'mode'.
If the value is 'test', the development mode with development filter
and production filter (if lavender index and lavender nodes are present)
is instanciated.